end work with converter

+++ src/models/Exercise.java
@@ -3,10 +3,10 @@ package models;
 public class Exercise {
     public final String name;
     public final String id;
-    public Integer points;
+    public Double points;
     public final String type;
 
-    public Exercise(String name, String id, Integer points, String type){
+    public Exercise(String name, String id, Double points, String type){
          this.name = name;
          this.id = id;
          this.points = points;
